Split renderer.js into focused modules (#133)
Extract the 3214-line renderer.js monolith into 6 focused modules + a slim orchestrator:

- renderer-state.js — shared mutable state, DOM refs, status classes, utilities
- editor.js — CodeMirror 6 live preview editor setup
- session-sidebar.js — session list rendering, directory colors, context menus, snapshots
- terminal-manager.js — terminal creation, attach, switch, close, caching, reconnect, PTY IPC
- pool-ui.js — pool settings panel, slot terminal popup, shortcut settings
- command-palette.js — COMMANDS registry, pane navigation, palette UI
- renderer.js — orchestrator (session lifecycle, auto-save, IPC wiring, module init)

Pure extraction refactoring — no behavioral changes. All 255 tests pass.
