+++ src/editor.js
@@ -1,0 +1,304 @@
+import { EditorState } from "@codemirror/state";
+import {
+  EditorView,
+  keymap,
+  Decoration,
+  ViewPlugin,
+  WidgetType,
+} from "@codemirror/view";
+import { markdown, markdownLanguage } from "@codemirror/lang-markdown";
+import { defaultKeymap, history, historyKeymap } from "@codemirror/commands";
+import { syntaxTree } from "@codemirror/language";
+import { state } from "./renderer-state.js";
+
+// --- Callback for doc changes (set by renderer.js) ---
+let _onDocChange = null;
+export function setOnDocChange(fn) {
+  _onDocChange = fn;
+}
+
+// --- Bullet widget: replaces "- " / "* " / "1. " with rendered bullet ---
+class BulletWidget extends WidgetType {
+  constructor(ordered, index) {
+    super();
+    this.ordered = ordered;
+    this.index = index;
+  }
+  toDOM() {
+    const span = document.createElement("span");
+    span.className = "cm-md-bullet-char";
+    span.textContent = this.ordered ? `${this.index}.  ` : "•  ";
+    return span;
+  }
+  ignoreEvent() {
+    return false;
+  }
+}
+
+// --- Live preview theme ---
+const livePreviewTheme = EditorView.theme({
+  "&": {
+    fontSize: "14px",
+    fontFamily:
+      "-apple-system, BlinkMacSystemFont, 'SF Pro Text', system-ui, sans-serif",
+  },
+  ".cm-content": {
+    padding: "16px 24px",
+    maxWidth: "800px",
+    lineHeight: "1.7",
+  },
+  ".cm-line": {
+    padding: "1px 0",
+  },
+  ".cm-md-heading1": {
+    fontSize: "1.8em",
+    fontWeight: "700",
+    color: "#ff1a1a",
+    marginTop: "8px",
+  },
+  ".cm-md-heading2": {
+    fontSize: "1.4em",
+    fontWeight: "600",
+    color: "#ff4444",
+    marginTop: "6px",
+  },
+  ".cm-md-heading3": {
+    fontSize: "1.2em",
+    fontWeight: "600",
+    color: "#ff6666",
+    marginTop: "4px",
+  },
+  ".cm-md-bold": { fontWeight: "700" },
+  ".cm-md-italic": { fontStyle: "italic" },
+  ".cm-md-code": {
+    fontFamily: "'SF Mono', Menlo, monospace",
+    fontSize: "13px",
+    background: "#252525",
+    padding: "1px 5px",
+    borderRadius: "3px",
+    color: "#ff6666",
+  },
+  ".cm-md-link": { color: "#ff4444", textDecoration: "underline" },
+  ".cm-md-strikethrough": { textDecoration: "line-through", color: "#808080" },
+  ".cm-md-bullet-char": {
+    color: "#ff1a1a",
+    fontWeight: "600",
+    marginRight: "2px",
+  },
+  ".cm-md-list-line": { paddingLeft: "8px" },
+  ".cm-md-blockquote": {
+    borderLeft: "3px solid #ff1a1a",
+    paddingLeft: "14px",
+    color: "#999999",
+  },
+  ".cm-md-hr": {
+    display: "block",
+    borderTop: "1px solid #252525",
+    margin: "12px 0",
+  },
+  ".cm-md-checkbox": { marginRight: "4px" },
+});
+
+const INLINE_STYLES = {
+  StrongEmphasis: "cm-md-bold",
+  Emphasis: "cm-md-italic",
+  InlineCode: "cm-md-code",
+  Strikethrough: "cm-md-strikethrough",
+};
+
+const SYNTAX_MARKS = new Set([
+  "HeaderMark",
+  "EmphasisMark",
+  "CodeMark",
+  "StrikethroughMark",
+  "QuoteMark",
+]);
+
+function getActiveLines(state) {
+  const lines = new Set();
+  for (const range of state.selection.ranges) {
+    const startLine = state.doc.lineAt(range.from).number;
+    const endLine = state.doc.lineAt(range.to).number;
+    for (let i = startLine; i <= endLine; i++) {
+      lines.add(i);
+    }
+  }
+  return lines;
+}
+
+function buildDecorations(view) {
+  const { state } = view;
+  const activeLines = getActiveLines(state);
+  const decorations = [];
+  const tree = syntaxTree(state);
+  const processedLines = new Set();
+  let orderedIndex = 0;
+  let linkDepth = 0;
+  const isLinkContainer = (n) => n === "Link" || n === "Image";
+
+  tree.iterate({
+    leave(node) {
+      if (isLinkContainer(node.name)) linkDepth--;
+    },
+    enter(node) {
+      if (isLinkContainer(node.name)) linkDepth++;
+      const line = state.doc.lineAt(node.from);
+      const isActive = activeLines.has(line.number);
+
+      if (node.name.startsWith("ATXHeading") && !node.name.includes("Mark")) {
+        const level = node.name.match(/(\d)$/)?.[1] || "1";
+        if (!processedLines.has(`heading-${line.number}`)) {
+          decorations.push(
+            Decoration.line({ class: `cm-md-heading${level}` }).range(
+              line.from,
+            ),
+          );
+          processedLines.add(`heading-${line.number}`);
+        }
+        if (isActive) return;
+      }
+
+      if (isActive) return;
+
+      if (node.name === "Blockquote") {
+        const startLine = state.doc.lineAt(node.from).number;
+        const endLine = state.doc.lineAt(node.to).number;
+        for (let i = startLine; i <= endLine; i++) {
+          if (!activeLines.has(i) && !processedLines.has(`bq-${i}`)) {
+            decorations.push(
+              Decoration.line({ class: "cm-md-blockquote" }).range(
+                state.doc.line(i).from,
+              ),
+            );
+            processedLines.add(`bq-${i}`);
+          }
+        }
+      }
+
+      if (node.name === "ListMark") {
+        const text = state.doc.sliceString(node.from, node.to);
+        const isOrdered = /^\d+[.)]$/.test(text);
+        if (isOrdered) {
+          orderedIndex++;
+        } else {
+          orderedIndex = 0;
+        }
+        let end = node.to;
+        if (
+          end < state.doc.length &&
+          state.doc.sliceString(end, end + 1) === " "
+        ) {
+          end += 1;
+        }
+        decorations.push(
+          Decoration.replace({
+            widget: new BulletWidget(isOrdered, orderedIndex),
+          }).range(node.from, end),
+        );
+        if (!processedLines.has(`list-${line.number}`)) {
+          decorations.push(
+            Decoration.line({ class: "cm-md-list-line" }).range(line.from),
+          );
+          processedLines.add(`list-${line.number}`);
+        }
+        return;
+      }
+
+      if (SYNTAX_MARKS.has(node.name) && node.from < node.to) {
+        decorations.push(Decoration.replace({}).range(node.from, node.to));
+      }
+      if (node.name === "LinkMark" && node.from < node.to) {
+        decorations.push(Decoration.replace({}).range(node.from, node.to));
+      }
+      // Only hide URL nodes inside Link/Image (e.g. [text](url)), not standalone bare URLs
+      if (node.name === "URL" && node.from < node.to && linkDepth > 0) {
+        decorations.push(Decoration.replace({}).range(node.from, node.to));
+      }
+
+      const styleClass = INLINE_STYLES[node.name];
+      if (styleClass && node.from < node.to) {
+        decorations.push(
+          Decoration.mark({ class: styleClass }).range(node.from, node.to),
+        );
+      }
+      if (
+        (node.name === "Link" ||
+          node.name === "Autolink" ||
+          (node.name === "URL" && linkDepth === 0)) &&
+        node.from < node.to
+      ) {
+        decorations.push(
+          Decoration.mark({ class: "cm-md-link" }).range(node.from, node.to),
+        );
+      }
+      if (node.name === "HorizontalRule") {
+        decorations.push(
+          Decoration.line({ class: "cm-md-hr" }).range(line.from),
+        );
+      }
+    },
+  });
+
+  decorations.sort((a, b) => a.from - b.from || a.startSide - b.startSide);
+  return Decoration.set(decorations, true);
+}
+
+const livePreviewPlugin = ViewPlugin.fromClass(
+  class {
+    constructor(view) {
+      this.decorations = buildDecorations(view);
+    }
+    update(update) {
+      if (update.docChanged || update.selectionSet || update.viewportChanged) {
+        this.decorations = buildDecorations(update.view);
+      }
+    }
+  },
+  { decorations: (v) => v.decorations },
+);
+
+const darkTheme = EditorView.theme(
+  {
+    "&": {
+      backgroundColor: "#0a0a0a",
+      color: "#e0e0e0",
+      height: "100%",
+    },
+    ".cm-cursor": { borderLeftColor: "#c0c0c0" },
+    ".cm-selectionBackground, &.cm-focused .cm-selectionBackground": {
+      backgroundColor: "#2a0808",
+    },
+    ".cm-gutters": { display: "none" },
+    ".cm-activeLine": { backgroundColor: "#121212" },
+    "&.cm-focused": { outline: "none" },
+    ".cm-scroller": { overflow: "auto" },
+  },
+  { dark: true },
+);
+
+function createEditor(content) {
+  if (state.editorView) state.editorView.destroy();
+
+  const editorState = EditorState.create({
+    doc: content,
+    extensions: [
+      history(),
+      keymap.of([...defaultKeymap, ...historyKeymap]),
+      markdown({ base: markdownLanguage }),
+      livePreviewPlugin,
+      livePreviewTheme,
+      darkTheme,
+      EditorView.lineWrapping,
+      EditorView.updateListener.of((update) => {
+        if (update.docChanged && _onDocChange) _onDocChange();
+      }),
+    ],
+  });
+
+  state.editorView = new EditorView({
+    state: editorState,
+    parent: state.editorMount,
+  });
+}
+
+export { createEditor };
